requirements.txt
flask>=2.2.2
flask-login>=0.6.2
google-api-python-client>=2.118.0
google-auth>=2.28.1
google-auth-oauthlib>=1.2.0
requests>=2.31.0
werkzeug>=2.2.2
itsdangerous>=2.1.2
jinja2>=3.1.2
gunicorn>=20.1.0
click>=8.1.3
markupsafe>=2.1.3
